fridge-scanner-service/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
